requirements.txt
django-registration
django-session-stashable
south
django-lazysignup
webtest
django-webtest
django-nose
